updated Header and app.jsx
